WMR918: Add init timeout of 60 seconds
If one of the sensors is not connected, proceed with initialization
after a timeout (currently 60 seconds).  The timeout is not ideal, but
due to variability of packet type frequency, there is no clear reliable
alternative.

This condition occurs intentionally when sensors are unused (e.g. RAIN
in snow climates) or unintentionally (e.g. due to malfunction).  In
either case, I think it's better better to gather some data with
user-obvious defects than no data with logged status messages
indefinitely.  When proposed, the upstream author contended that users
must ensure all sensors are functioning for wview to work, so this patch
is unlikely to be merged.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>
Bug: http://www.wviewweather.com/mantis/view.php?id=126