WMR918: reopen station device if it disappears
If the station disconnects (e.g. due to power failure) reads will return
ENOENT indefinitely.  When this happens, reopen the device (restart the
medium).

This patch is not as elegant as I would like in terms of code (e.g. the
goto, which was clearer than loop/break used as goto) and in terms of
API (readStationData doesn't preserve errno or propagate the error cause
by another mechanism, so it has to be handled here).  But it works for
me and that's all I need for now.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>