Install Apache configuration for wview
Install Apache configuration files with aliases and permissions for
/weather and /wviewmgmt (as before) using dh-apache2.

This is the conventional approach for packages which provide a
web-accessible endpoint (see gitweb, javascript-common, mediawiki,
w3c-markup-validator, etc.) and is in compliance with the Debian Webapps
Policy and apache2 PACKAGING doc.

It also avoids the need for linking from guessed ServerRoot directories
in postinst, which this commit removes.

Note:  Introduces lintian error due to wviewmgmt config not matching
package name.  This is intentional to allow enabling/disabling the
management website separately from the data website.  It may also be
fixed if wviewmgmt is split into a separate package.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>