Don't install functions.php as a binary
Adding functions.php to bin_SCRIPTS causes it to be installed as an
executable script in /usr/bin.  functions.php is a generated PHP file
and not an executable that can/should be invoked.  Remove from
bin_SCRIPTS, add to EXTRA_DIST, and don't link to the copy in /usr/bin.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>