Create wview user for daemons, set file perms
Run the daemons as an unprivileged user created by this package and set
restrictive permissions on files installed by this package in order to
reduce the attack surface and potential for privilege escalation.

Note:  After considering the [arguments for and against user removal[1],
I have chosen not to remove the user on uninstall due to the amount and
variety of data (raw sqlite data, generated HTML or other templated
data, and any export/upload data) which is likely to remain in backups
and various remote and/or local system locations.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>