Set correct execute permission on files
Giving execute permission to files which can not be executed is
misleading, risks security issues (e.g. if the data can be changed and
something tricked in to executing it), and causes the Lintian
executable-not-elf-or-script and script-not-executable warnings.  Fix
these by marking executable files (binaries and scripts with a shebang)
as executable and all other files as non-executable.

Note:  Skip files with spaces in name due to quilt failing to apply the
exported patch.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>